refactor: extract function to its own file in the utils directory
